requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --annotation-style=line --strip-extras requirements.in
#
annotated-types==0.7.0    # via pydantic
apscheduler==3.11.0       # via -r requirements.in
boto3==1.35.72            # via -r requirements.in
botocore==1.35.72         # via boto3, s3transfer
jmespath==1.0.1           # via boto3, botocore
pydantic==2.10.2          # via -r requirements.in
pydantic-core==2.27.1     # via pydantic
python-dateutil==2.9.0.post0  # via botocore
pyyaml==6.0.2             # via -r requirements.in
s3transfer==0.10.4        # via boto3
six==1.16.0               # via python-dateutil
tenacity==9.0.0           # via -r requirements.in
typing-extensions==4.12.2  # via pydantic, pydantic-core
tzlocal==5.2              # via apscheduler
urllib3==2.2.3            # via botocore
